Fix bug in themes and permissioning

+++ run-ghost.sh
@@ -1,13 +1,11 @@
 #!/bin/bash
 
-if ! test -e var/ghostcms; then
-    mkdir -p var/ghostcms
-    mkdir -p var/static
+test -e var/ghostcms || mkdir -p var/ghostcms
+mkdir -p var/www
 
-    cp -R content var/ghostcms
-    rm -rf var/ghostcms/content/themes
-    ln -s /themes var/ghostcms/content/themes
-    cp dbseed/*.db  var/ghostcms/content/data
-fi
+test -e var/ghostcms/content || cp -R content var/ghostcms
+rm -rf var/ghostcms/content/themes
+ln -s /themes var/ghostcms/content/themes
+test -e var/ghostcms/content/data/ghost-dev.db || cp dbseed/ghost-dev.db  var/ghostcms/content/data
 
 node index
